+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for Polar heart rate sensor, reading hear rate and RR-intervals,
   see [HRS_SPEC_v10](https://www.bluetooth.org/docman/handlers/downloaddoc.ashx?doc_id=239866).
 - Support for saving measurements into CSV file.
+- WIP: support for OxyOxySmart SpO2 sensor, not parsing the actual values yet.